refactor(api-core): move AdminGpuReset out of endpoint explorer, restrict to ForceRestart
Address review feedback:
- Move admin_gpu_reset (and its action mapping + test) into a dedicated
  handlers/gpu_reset.rs module instead of bmc_endpoint_explorer, since a
  break-fix chassis reset is an admin action, not endpoint exploration.
- Restrict the reset action to ForceRestart for v1. Redfish Chassis.Reset
  allowable ResetType values are vendor-specific and do not reliably include
  GracefulRestart or ACPowercycle, so reject those (and power-off actions)
  rather than send an unsupported ResetType to the BMC.
- Regenerate Core proto bindings and update the RPC docs accordingly.
